+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Rejects all errors with `FetchError`.
+
 ## 0.2.0 (Dec 9, 2015)
 - Rejects with `FetchError` instead of `HttpError`.
 - Adds the `request` property (an instance of [`Request`][]) with details of the